Explore: Set display processor for table data frame
