+++ packages/grafana-ui/src/components/Table/Table.tsx
@@ -1,324 +1,99 @@
-// Libraries
-import _ from 'lodash';
-import React, { Component, ReactElement } from 'react';
-import {
-  SortDirectionType,
-  SortIndicator,
-  MultiGrid,
-  CellMeasurerCache,
-  CellMeasurer,
-  GridCellProps,
-  Index,
-} from 'react-virtualized';
-import { Themeable } from '../../types/theme';
+import React, { useMemo, CSSProperties } from 'react';
+import { DataFrame } from '@grafana/data';
+// @ts-ignore
+import { useSortBy, useTable, useBlockLayout } from 'react-table';
+import { FixedSizeList } from 'react-window';
+import { getTableStyles } from './styles';
+import { getColumns, getTableRows, TableColumn } from './models';
+import { useTheme } from '../../themes';
 
-import {
-  stringToJsRegex,
-  DataFrame,
-  sortDataFrame,
-  getDataFrameRow,
-  ArrayVector,
-  FieldType,
-  InterpolateFunction,
-} from '@grafana/data';
-
-import {
-  TableCellBuilder,
-  ColumnStyle,
-  getFieldCellBuilder,
-  TableCellBuilderOptions,
-  simpleCellBuilder,
-} from './TableCellBuilder';
-
-export interface Props extends Themeable {
+export interface Props {
   data: DataFrame;
-
-  minColumnWidth: number;
-  showHeader: boolean;
-  fixedHeader: boolean;
-  fixedColumns: number;
-  rotate: boolean;
-  styles: ColumnStyle[];
-
-  replaceVariables: InterpolateFunction;
   width: number;
   height: number;
-  isUTC?: boolean;
+  onCellClick?: (key: string, value: string) => void;
 }
 
-interface State {
-  sortBy?: number;
-  sortDirection?: SortDirectionType;
-  data: DataFrame;
+interface RenderCellProps {
+  column: TableColumn;
+  value: any;
+  getCellProps: () => { style: CSSProperties };
+  render: (component: string) => React.ReactNode;
 }
 
-interface ColumnRenderInfo {
-  header: string;
-  width: number;
-  builder: TableCellBuilder;
-}
+function renderCell(cell: RenderCellProps, className: string, onCellClick?: any) {
+  const filterable = cell.column.field.config.filterable;
+  const cellProps = cell.getCellProps();
+  let onClick: ((event: React.SyntheticEvent) => void) | undefined = undefined;
 
-interface DataIndex {
-  column: number;
-  row: number; // -1 is the header!
-}
-
-export class Table extends Component<Props, State> {
-  renderer: ColumnRenderInfo[];
-  measurer: CellMeasurerCache;
-  scrollToTop = false;
-  rotateWidth = 100;
-
-  static defaultProps = {
-    showHeader: true,
-    fixedHeader: true,
-    fixedColumns: 0,
-    rotate: false,
-    minColumnWidth: 150,
-  };
-
-  constructor(props: Props) {
-    super(props);
-
-    this.state = {
-      data: props.data,
-    };
-
-    this.renderer = this.initColumns(props);
-    this.measurer = new CellMeasurerCache({
-      defaultHeight: 30,
-      fixedWidth: true,
-    });
+  if (filterable && onCellClick) {
+    cellProps.style.cursor = 'pointer';
+    onClick = () => onCellClick(cell.column.Header, cell.value);
   }
 
-  componentDidUpdate(prevProps: Props, prevState: State) {
-    const { data, styles, showHeader, rotate } = this.props;
-    const { sortBy, sortDirection } = this.state;
-    const dataChanged = data !== prevProps.data;
-    const configsChanged =
-      showHeader !== prevProps.showHeader ||
-      this.props.rotate !== prevProps.rotate ||
-      this.props.fixedColumns !== prevProps.fixedColumns ||
-      this.props.fixedHeader !== prevProps.fixedHeader;
-
-    // Reset the size cache
-    if (dataChanged || configsChanged) {
-      this.measurer.clearAll();
-    }
-
-    // Update the renderer if options change
-    // We only *need* do to this if the header values changes, but this does every data update
-    if (dataChanged || styles !== prevProps.styles) {
-      this.renderer = this.initColumns(this.props);
-    }
-
-    if (dataChanged || rotate !== prevProps.rotate) {
-      const { width, minColumnWidth } = this.props;
-      this.rotateWidth = Math.max(width / data.length, minColumnWidth);
-    }
-
-    // Update the data when data or sort changes
-    if (dataChanged || sortBy !== prevState.sortBy || sortDirection !== prevState.sortDirection) {
-      this.scrollToTop = true;
-      this.setState({ data: sortDataFrame(data, sortBy, sortDirection === 'DESC') });
-    }
+  if (cell.column.textAlign) {
+    cellProps.style.textAlign = cell.column.textAlign;
   }
 
-  /** Given the configuration, setup how each column gets rendered */
-  initColumns(props: Props): ColumnRenderInfo[] {
-    const { styles, data, width, minColumnWidth } = props;
-    if (!data || !data.fields || !data.fields.length || !styles) {
-      return [];
-    }
+  return (
+    <div className={className} {...cellProps} onClick={onClick}>
+      {cell.render('Cell')}
+    </div>
+  );
+}
 
-    const columnWidth = Math.max(width / data.fields.length, minColumnWidth);
+function renderHeader(column: any, className: string) {
+  const headerProps = column.getHeaderProps(column.getSortByToggleProps());
 
-    return data.fields.map((col, index) => {
-      let title = col.name;
-      let style: ColumnStyle | null = null; // ColumnStyle
-
-      // Find the style based on the text
-      for (let i = 0; i < styles.length; i++) {
-        const s = styles[i];
-        const regex = stringToJsRegex(s.pattern);
-        if (title.match(regex)) {
-          style = s;
-          if (s.alias) {
-            title = title.replace(regex, s.alias);
-          }
-          break;
-        }
-      }
-
-      return {
-        header: title,
-        width: columnWidth,
-        builder: getFieldCellBuilder(col, style, this.props),
-      };
-    });
+  if (column.textAlign) {
+    headerProps.style.textAlign = column.textAlign;
   }
 
-  //----------------------------------------------------------------------
-  //----------------------------------------------------------------------
+  return (
+    <div className={className} {...headerProps}>
+      {column.render('Header')}
+      <span>{column.isSorted ? (column.isSortedDesc ? ' 🔽' : ' 🔼') : ''}</span>
+    </div>
+  );
+}
 
-  doSort = (columnIndex: number) => {
-    let sort: any = this.state.sortBy;
-    let dir = this.state.sortDirection;
-    if (sort !== columnIndex) {
-      dir = 'DESC';
-      sort = columnIndex;
-    } else if (dir === 'DESC') {
-      dir = 'ASC';
-    } else {
-      sort = null;
-    }
-    this.setState({ sortBy: sort, sortDirection: dir });
-  };
+export const Table = ({ data, height, onCellClick, width }: Props) => {
+  const theme = useTheme();
+  const tableStyles = getTableStyles(theme);
+  const { getTableProps, headerGroups, rows, prepareRow } = useTable(
+    {
+      columns: useMemo(() => getColumns(data, width, theme), [data]),
+      data: useMemo(() => getTableRows(data), [data]),
+    },
+    useSortBy,
+    useBlockLayout
+  );
 
-  /** Converts the grid coordinates to DataFrame coordinates */
-  getCellRef = (rowIndex: number, columnIndex: number): DataIndex => {
-    const { showHeader, rotate } = this.props;
-    const rowOffset = showHeader ? -1 : 0;
+  const RenderRow = React.useCallback(
+    ({ index, style }) => {
+      const row = rows[index];
+      prepareRow(row);
+      return (
+        <div {...row.getRowProps({ style })}>
+          {row.cells.map((cell: RenderCellProps) => renderCell(cell, tableStyles.tableCell, onCellClick))}
+        </div>
+      );
+    },
+    [prepareRow, rows]
+  );
 
-    if (rotate) {
-      return { column: rowIndex, row: columnIndex + rowOffset };
-    } else {
-      return { column: columnIndex, row: rowIndex + rowOffset };
-    }
-  };
-
-  onCellClick = (rowIndex: number, columnIndex: number) => {
-    const { row, column } = this.getCellRef(rowIndex, columnIndex);
-    if (row < 0) {
-      this.doSort(column);
-    } else {
-      const field = this.state.data.fields[columnIndex];
-      const value = field.values.get(rowIndex);
-      console.log('CLICK', value, field.name);
-    }
-  };
-
-  headerBuilder = (cell: TableCellBuilderOptions): ReactElement<'div'> => {
-    const { data, sortBy, sortDirection } = this.state;
-    const { columnIndex, rowIndex, style } = cell.props;
-    const { column } = this.getCellRef(rowIndex, columnIndex);
-
-    let col = data.fields[column];
-    const sorting = sortBy === column;
-    if (!col) {
-      col = {
-        name: '??' + columnIndex + '???',
-        config: {},
-        values: new ArrayVector(),
-        type: FieldType.other,
-      };
-    }
-
-    return (
-      <div className="gf-table-header" style={style} onClick={() => this.onCellClick(rowIndex, columnIndex)}>
-        {col.name}
-        {sorting && <SortIndicator sortDirection={sortDirection} />}
+  return (
+    <div {...getTableProps()} className={tableStyles.table}>
+      <div>
+        {headerGroups.map((headerGroup: any) => (
+          <div className={tableStyles.thead} {...headerGroup.getHeaderGroupProps()}>
+            {headerGroup.headers.map((column: any) => renderHeader(column, tableStyles.headerCell))}
+          </div>
+        ))}
       </div>
-    );
-  };
-
-  getTableCellBuilder = (column: number): TableCellBuilder => {
-    const render = this.renderer[column];
-    if (render && render.builder) {
-      return render.builder;
-    }
-    return simpleCellBuilder; // the default
-  };
-
-  cellRenderer = (props: GridCellProps): React.ReactNode => {
-    const { rowIndex, columnIndex, key, parent } = props;
-    const { row, column } = this.getCellRef(rowIndex, columnIndex);
-    const { data } = this.state;
-
-    const isHeader = row < 0;
-    const rowData = isHeader ? data.fields : getDataFrameRow(data, row); // TODO! improve
-    const value = rowData ? rowData[column] : '';
-    const builder = isHeader ? this.headerBuilder : this.getTableCellBuilder(column);
-
-    return (
-      <CellMeasurer cache={this.measurer} columnIndex={columnIndex} key={key} parent={parent} rowIndex={rowIndex}>
-        {builder({
-          value,
-          row: rowData,
-          column: data.fields[column],
-          table: this,
-          props,
-        })}
-      </CellMeasurer>
-    );
-  };
-
-  getColumnWidth = (col: Index): number => {
-    if (this.props.rotate) {
-      return this.rotateWidth; // fixed for now
-    }
-    return this.renderer[col.index].width;
-  };
-
-  render() {
-    const { showHeader, fixedHeader, fixedColumns, rotate, width, height } = this.props;
-    const { data } = this.state;
-    if (!data || !data.fields || !data.fields.length) {
-      return <span>Missing Fields</span>; // nothing
-    }
-
-    let columnCount = data.fields.length;
-    let rowCount = data.length + (showHeader ? 1 : 0);
-
-    let fixedColumnCount = Math.min(fixedColumns, columnCount);
-    let fixedRowCount = showHeader && fixedHeader ? 1 : 0;
-
-    if (rotate) {
-      const temp = columnCount;
-      columnCount = rowCount;
-      rowCount = temp;
-
-      fixedRowCount = 0;
-      fixedColumnCount = Math.min(fixedColumns, rowCount) + (showHeader && fixedHeader ? 1 : 0);
-    }
-
-    // Called after sort or the data changes
-    const scroll = this.scrollToTop ? 1 : -1;
-    const scrollToRow = rotate ? -1 : scroll;
-    const scrollToColumn = rotate ? scroll : -1;
-    if (this.scrollToTop) {
-      this.scrollToTop = false;
-    }
-
-    // Force MultiGrid to rerender if these options change
-    // See: https://github.com/bvaughn/react-virtualized#pass-thru-props
-    const refreshKeys = {
-      ...this.state, // Includes data and sort parameters
-      d1: this.props.data,
-      s0: this.props.styles,
-    };
-    return (
-      <MultiGrid
-        {...refreshKeys}
-        scrollToRow={scrollToRow}
-        columnCount={columnCount}
-        scrollToColumn={scrollToColumn}
-        rowCount={rowCount}
-        overscanColumnCount={8}
-        overscanRowCount={8}
-        columnWidth={this.getColumnWidth}
-        deferredMeasurementCache={this.measurer}
-        cellRenderer={this.cellRenderer}
-        rowHeight={this.measurer.rowHeight}
-        width={width}
-        height={height}
-        fixedColumnCount={fixedColumnCount}
-        fixedRowCount={fixedRowCount}
-        classNameTopLeftGrid="gf-table-fixed-column"
-        classNameBottomLeftGrid="gf-table-fixed-column"
-      />
-    );
-  }
-}
-
-export default Table;
+      <FixedSizeList height={height} itemCount={rows.length} itemSize={tableStyles.cellHeight} width={width}>
+        {RenderRow}
+      </FixedSizeList>
+    </div>
+  );
+};
