+++ packages/grafana-ui/src/components/index.scss
@@ -9,7 +9,7 @@
 @import 'PanelOptionsGroup/PanelOptionsGroup';
 @import 'RefreshPicker/RefreshPicker';
 @import 'Select/Select';
-@import 'Table/TableInputCSV';
+@import 'TableInputCSV/TableInputCSV';
 @import 'ThresholdsEditor/ThresholdsEditor';
 @import 'TimePicker/TimeOfDayPicker';
 @import 'Tooltip/Tooltip';
