+++ public/app/features/explore/utils/ResultProcessor.ts
@@ -1,9 +1,18 @@
-import { LogsModel, GraphSeriesXY, DataFrame, FieldType, TimeZone, toDataFrame } from '@grafana/data';
+import {
+  LogsModel,
+  GraphSeriesXY,
+  DataFrame,
+  FieldType,
+  TimeZone,
+  toDataFrame,
+  getDisplayProcessor,
+} from '@grafana/data';
 import { ExploreItemState, ExploreMode } from 'app/types/explore';
 import TableModel, { mergeTablesIntoModel } from 'app/core/table_model';
 import { sortLogsResult, refreshIntervalToSortOrder } from 'app/core/utils/explore';
 import { dataFrameToLogsModel } from 'app/core/logs_model';
 import { getGraphSeriesModel } from 'app/plugins/panel/graph2/getGraphSeriesModel';
+import { config } from 'app/core/config';
 
 export class ResultProcessor {
   constructor(
@@ -75,7 +84,17 @@ export class ResultProcessor {
     });
 
     const mergedTable = mergeTablesIntoModel(new TableModel(), ...tables);
-    return toDataFrame(mergedTable);
+    const data = toDataFrame(mergedTable);
+
+    // set display processor
+    for (const field of data.fields) {
+      field.display = getDisplayProcessor({
+        config: field.config,
+        theme: config.theme,
+      });
+    }
+
+    return data;
   }
 
   getLogsResult(): LogsModel | null {
