+++ public/app/plugins/panel/table/types.ts
@@ -1,6 +1,5 @@
 import TableModel from 'app/core/table_model';
 import { Column } from '@grafana/data';
-import { ColumnStyle } from '@grafana/ui/src/components/Table/TableCellBuilder';
 
 export interface TableTransform {
   description: string;
@@ -17,4 +16,27 @@ export interface ColumnRender extends Column {
 export interface TableRenderModel {
   columns: ColumnRender[];
   rows: any[][];
+}
+
+export interface ColumnStyle {
+  pattern: string;
+
+  alias?: string;
+  colorMode?: 'cell' | 'value';
+  colors?: any[];
+  decimals?: number;
+  thresholds?: any[];
+  type?: 'date' | 'number' | 'string' | 'hidden';
+  unit?: string;
+  dateFormat?: string;
+  sanitize?: boolean; // not used in react
+  mappingType?: any;
+  valueMaps?: any;
+  rangeMaps?: any;
+
+  link?: any;
+  linkUrl?: any;
+  linkTooltip?: any;
+  linkTargetBlank?: boolean;
+  preserveFormat?: boolean;
 }
